Move library structs into the dictionary
The main file was getting a bit cluttered so I moved the contents into
their own module (`dictionary`). This module contains the structs with
some changes to conform better with the actual use case of the pieces. A
sha is broken into "parts" which then can be looked up for "strings"
instead of words. This seems more semantic and avoids modifying the word
to have a word after passing the sha in.

This is also the first step towards busting it out into a separate
crate.
