+++ src/dictionary/sha_result.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Eq, PartialEq)]
+pub enum ParseShaError {
+    NonHexadecimalCharacters,
+    WordNotFound,
+}
+
+pub type ShaResult<T> = Result<T, ParseShaError>;
